requirements.txt
django~=3.1.2
django-rest-framework~=0.1.0
psycopg2~=2.8.6
celery~=4.4.7
redis~=3.5.3
pylint~=2.6.0
pylint-django~=2.3.0
